# docs/openapi.yml
openapi: "3.0.3"
info:
  title: Medusa API
  version: 1.0.0
  description: E-commerce backend API

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 6
                  example: securePass123
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "registered successfully"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"

  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                token: eyJhbGciOiJIUzI1NiIs...
                refreshToken: rtok_019eca3c-...
                metadata:
                  requestId: "019eca3c-..."
                  timestamp: "2026-06-15T07:43:58Z"
                  latencyUs: 83390
        "400":
          $ref: "#/components/responses/ValidationError"

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
              example:
                token: eyJhbGciOiJIUzI1NiIs...
                refreshToken: rtok_019eca3c-...
        "400":
          $ref: "#/components/responses/ValidationError"

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh tokens
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: refresh_token
          schema:
            type: string
          description: Optional specific refresh token to revoke
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              example:
                message: logged out
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  /customers/me:
    get:
      tags: [Customer]
      summary: Get authenticated customer profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    $ref: "#/components/schemas/Customer"
              example:
                customer:
                  id: cus_019eca3c-...
                  email: john@example.com
                  firstName: John
                  lastName: Doe
                  phone: null
                  companyName: null
                  hasAccount: true
                  createdAt: "2026-06-15T07:43:58Z"
                  updatedAt: "2026-06-15T07:43:58Z"
                  deletedAt: null
        "401":
          $ref: "#/components/responses/UnauthorizedError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token (30min expiry)
        refreshToken:
          type: string
          description: Refresh token (30 day expiry)

    RefreshResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string

    Customer:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        hasAccount:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"